database/requirements.txt
# Firebase Admin SDK for Firestore access
firebase-admin>=6.5.0

# Data processing
pandas>=2.2.0

# CLI framework
click>=8.1.7

# Data validation
pydantic>=2.6.0

# Progress bars
tqdm>=4.66.0

# Date handling
python-dateutil>=2.8.2

# Logging
colorlog>=6.8.0

# Google Cloud SQL connector (for MySQL sync)
cloud-sql-python-connector[pymysql]>=1.10.0
pymysql>=1.1.0
google-auth>=2.28.0
